id to int
id to int

+++ carRentForCompany2/app/src/main/java/com/example/moish/carrentforcompany/model/entities/Client.java
@@ -9,7 +9,7 @@ public class Client {
 
     String lastName;
     String firstName;
-    long id;
+    int id;
     String phoneNumber;
     String email;
     int creditCardNumber;
